+++ client/Socket.h
@@ -148,9 +148,9 @@ public:
 	virtual void connect(const string& aIp, short aPort) throw(SocketException);
 	void connect(const string& aIp, const string& aPort) throw(SocketException) { connect(aIp, (short)Util::toInt(aPort)); };
 	virtual void accept(const ServerSocket& aSocket) throw(SocketException);
-	virtual void write(const char* buffer, int len) throw(SocketException);
+	virtual void write(const char* buffer, size_t len) throw(SocketException);
 	void write(const string& aData) throw(SocketException) { write(aData.data(), aData.length()); };
-	virtual void writeTo(const string& aIp, short aPort, const char* buffer, int len) throw(SocketException);
+	virtual void writeTo(const string& aIp, short aPort, const char* buffer, size_t len) throw(SocketException);
 	void writeTo(const string& aIp, short aPort, const string& aData) throw(SocketException) { writeTo(aIp, aPort, aData.data(), aData.length()); };
 
 	int read(void* aBuffer, int aBufLen) throw(SocketException);
@@ -258,6 +258,6 @@ private:
 
 /**
  * @file
- * $Id: Socket.h,v 1.54 2004/09/06 12:32:42 arnetheduck Exp $
+ * $Id: Socket.h,v 1.55 2004/09/09 09:27:36 arnetheduck Exp $
  */
