+++ client/TigerHash.h
@@ -38,7 +38,7 @@ public:
 	}
 
 	/** Calculates the Tiger hash of the data. */
-	void update(const void* data, u_int32_t len);
+	void update(const void* data, size_t len);
 	/** Call once all data has been processed. */
 	u_int8_t* finalize();
 
@@ -61,5 +61,5 @@ private:
 
 /**
  * @file
- * $Id: TigerHash.h,v 1.5 2004/09/06 12:32:43 arnetheduck Exp $
+ * $Id: TigerHash.h,v 1.6 2004/09/09 09:27:36 arnetheduck Exp $
  */
