+++ client/Util.cpp
@@ -61,7 +61,7 @@ static void sgenrand(unsigned long seed);
 void Util::initialize() {
 	setlocale(LC_ALL, "");
 
-	int i;
+	size_t i;
 	for(i = 0; i < 65536; ++i) {
 #ifdef _WIN32
 		lower[i] = (wchar_t)CharLowerW((LPWSTR)i);
@@ -80,7 +80,7 @@ void Util::initialize() {
 		}
 	}
 
-	sgenrand(time(NULL));
+	sgenrand((unsigned long)time(NULL));
 
 	try {
 		string file = Util::getAppPath() + "GeoIpCountryWhois.csv";
@@ -534,8 +534,8 @@ string& Util::toUtf8(string& str) {
 		return str;
 	wstring wtmp(str.length(), 0);
 #ifdef _WIN32
-	int sz = MultiByteToWideChar(CP_ACP, MB_PRECOMPOSED, str.c_str(), str.length(),
-		&wtmp[0], str.length());
+	int sz = MultiByteToWideChar(CP_ACP, MB_PRECOMPOSED, str.c_str(), (int)str.length(),
+		&wtmp[0], (int)str.length());
 	if(sz <= 0) {
 		str.clear();
 		return str;
@@ -579,13 +579,13 @@ string& Util::toAcp(string& str) {
 	utf8ToWide(str, wtmp);
 
 #ifdef _WIN32
-	int x = WideCharToMultiByte(CP_ACP, 0, wtmp.c_str(), wtmp.length(), NULL, 0, NULL, NULL);
+	int x = WideCharToMultiByte(CP_ACP, 0, wtmp.c_str(), (int)wtmp.length(), NULL, 0, NULL, NULL);
 	if(x == 0) {
 		str.clear();
 		return str;
 	}
 	str.resize(x);
-	WideCharToMultiByte(CP_ACP, 0, wtmp.c_str(), wtmp.length(), &str[0], str.size(), NULL, NULL);
+	WideCharToMultiByte(CP_ACP, 0, wtmp.c_str(), (int)wtmp.length(), &str[0], (int)str.size(), NULL, NULL);
 #else
 	size_t x = wcstombs(NULL, wtmp.c_str(), 0);
 	if(x == (size_t)-1) {
@@ -912,6 +912,6 @@ int Util::getOsMinor()
 }
 /**
  * @file
- * $Id: Util.cpp,v 1.62 2004/09/07 01:36:52 arnetheduck Exp $
+ * $Id: Util.cpp,v 1.63 2004/09/09 09:27:36 arnetheduck Exp $
  */
